move work to a script
